+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - feat: Add usage of Geometric Kandel's call-data-reducing function
 - feat!: Removed `configuration.tokens.fetchDecimalsFromAddress`. Instead, use `MgvToken.createTokenFromAddress` and read the decimals from that token.
 - feat!: `Mangrove.toUnits|fromUnits` no longer accepts a token name/symbol as this was ambiguous. Instead, use `MgvToken.createToken` and call `toUnits|fromUnits` on that.
+- feat!: Static token configuration getters and setters have been removed from `Mangrove`. Instead, use the methods on `MgvToken`.
 
 # 2.0.0-4